Add further optional arguments to `group_words'.
Distinguish between removing leading and trailing whitespace.  Allow
chunks to be up to `tol' characters longer than `n'.  Allow "punctuation
words", that is, words consisting of a single punctuation mark, to be
merged to the previous chunk irrespective of the chunk's length.  This
is only useful when splitting strings into a list of their characters.